Move a2l1c-specific functions to util module in fetch dir
This will help with the transition to using `pm_tb_data`
